Moves types to types.ts. Moves randomPositionAvoidingCenter function to utils.ts file on the same folder.
